pyproject.toml
[tool.poetry]
name = "coursework3"
version = "0.1.0"
description = ""
authors = ["Dmitriy Kurilko <dimakurilko0606@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
xlrd = "^2.0.1"
python-dotenv = "^1.0.1"
yfinance = "^0.2.40"
pytest = "^8.2.2"
freezegun = "^1.5.1"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
mypy = "^1.10.1"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]

# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''


[tool.isort]
# максимальная длина строки
line_length = 119


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
